app/requirements-dev.txt
-r requirements.txt
flake8==7.1.1
pytest==8.3.3
pytest-flask==1.3.0
